Renamed cms.templatetags.js to cms.templatetags.cms_js_tags (it's an internal library) Added cms.templatetags.cms_js_tags.JavascriptString (internal) templatetag Escape i18n strings passed to the toolbar in /cms/templates/cms/toolbar/toolbar.html using that new template tag
